+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     There's currently no way to use the modified database directly.
 - Changed: Removed Space Jump method from Training Chamber.
 - Changed: Added Power Bomb as option for pickup in Hive Chamber B.
+- Changed: Shortened Permalinks when pickup quantities aren't customized.
 
 ## [0.17.2] - 2018-12-27
 - Fixed: 'Clear loaded game' now properly does it's job.